+++ CVMento/src/main/java/com/cvmento/domain/resume/dto/response/EducationResponse.java
@@ -1,11 +1,13 @@
 package com.cvmento.domain.resume.dto.response;
 
+import java.math.BigDecimal;
+
 public record EducationResponse(
         String schoolName,
         String major,
         String degreeLevel,
-        Double personalGpa,
-        Double totalGpa,
+        BigDecimal personalGpa,
+        BigDecimal totalGpa,
         String graduationDate
 ) {
 }
